添加歌词转换器工具，支持ksc、krc和hrc歌词间的相互转换及转换成lrc歌词

+++ app/src/main/java/com/zlm/hp/ui/LrcActivity.java
@@ -1726,7 +1726,7 @@ public class LrcActivity extends BaseActivity {
      * @param lyricsInfo  lrc歌词数据
      * @param playOffset  lrc歌词快进进度
      */
-    private void saveLrcFile(final String lrcFilePath, final LyricsInfo lyricsInfo, final int playOffset) {
+    private void saveLrcFile(final String lrcFilePath, final LyricsInfo lyricsInfo, final long playOffset) {
         new Thread() {
 
             @Override
